fix: no subscription for default key (#187)
Follow-up to #163. I didn't remove all occurrences of `!!key`. For
rationale, please read #163. I did a write-up on what's going on here.

This time, this line prevents Rise from subscribing to `''` key. As a
result, if you change contents of the default screen on the server, it
won't refresh on the client unless you reload your screen.

After this PR, working with `''` is same as working with any other
screen.