feat: turn `example` into something people can start with (#198)
This PR makes it possible to get started with the full example as well.

I have removed simple-dev-server (counter example). Shall we merge the
counter example with the rest, that is, create a separate folder for it
and list it from `home`? I think it might not be necessary, given that
we're adding "vibes" in next PR, that already demonstrate state.

Alternatively, we could put Counter in UI Controls.

As a follow-up, I will make it so that you can choose features from the
full example, in the following manner:

- Get list of folders in `template/full/src` and treat each one of them
as a separate part of example
- List them all in the CLI together with explanation what they do
- Let users choose
- Remove folders that were not selected by the user
- Remove imports that are not needed in the server entry point