feat: improves DraggableFlatList (#213)
- Use TouchableWithoutFeedback to mitigate the issue where the pressed
state is stuck
- onPress callback for each item
- onPress callback for the whole component

Note: The issue is still present where the pressed state gets stuck down
after dragging an item. This is a bug because the user needs to press it
twice for the onPress callback to fire. But at least the user is not
left with a bad visual state. Obviously we want a real fix for this, and
when we have one, we can switch back to TouchableOpacity