feat: allow rendering iterable elements on the client (#215)
Following up to previous PR, this adds ability to render Iterables, not
just Arrays, as children and props. While our transportation mechanism
currently does not support transporting anything but arrays, this makes
it aligned with what server outputs, in case it gets passed directly as
a prop.

Previously, this would throw an error that objects can't be rendered,
which is correct, just not for iterables!