feat: add expo-video (#223)
This PR adds a new package that we will use for our conference demo.

The newly created <Video /> component contains props of VideoView from
`expo-view` as well as certain props from `useVideoSource` hook /
Player, that we set up for the user behind the scenes. That way, you use
a single component to control that behavior. Will add documentation
explaining this in a bit!

We currently do not support custom actions - this requires an
improvement to actions() to access refs/context relative to the
component that triggered action. We will likely have to look into this
sooner than later.

Screenshots:

| Screen    | Screen | Screen |
| -------- | ------- | ------ |
| ![Simulator Screenshot - iPhone 15 Pro - 2024-08-20 at 16 15
14](https://github.com/user-attachments/assets/1a3222e4-6d2b-4a3d-a57c-f4740a37d8d9)
| ![Simulator Screenshot - iPhone 15 Pro - 2024-08-20 at 16 15
25](https://github.com/user-attachments/assets/a505a46b-8754-4f82-afbd-f3fb936ea956)
| ![Simulator Screenshot - iPhone 15 Pro - 2024-08-20 at 16 15
22](https://github.com/user-attachments/assets/0bd91c0e-35e0-4979-b53d-b9248d3d7e8a)