refactor: remove dead Fragment code (#228)
JSX 
```
const View = (
  <>
    <div />
  </>
)
```
is transformed to this
```
import { jsx as _jsx, Fragment as _Fragment } from "react/jsx-runtime";
const View = /*#__PURE__*/_jsx(_Fragment, {
  children: /*#__PURE__*/_jsx("div", {})
});
```
As you can see, `Fragment` from `jsx-runtime` is used as
`componentFactory`, just like any other component. Beforehand, we didn't
have it, so it was undefined.

We actually already export it from `jsx-runtime`, so I just added it to
`jsx-dev-runtime` too to make it work and that `if` is no longer needed.